requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
python-dateutil==2.8.2
